Repaired potential integer overflow in LevelSet.getTileNumber(). See WWJ-451.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1862 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWind/src/gov/nasa/worldwind/util/LevelSet.java
@@ -359,13 +359,13 @@ public class LevelSet extends WWObjectImpl
 
     private long getTileNumber(Tile tile)
     {
-        return tile.getRow() < 0 ? -1 : tile.getRow() * this.numColumnsInLevel(tile.getLevel()) + tile.getColumn();
+        return tile.getRow() < 0 ? -1 : (long) tile.getRow() * this.numColumnsInLevel(tile.getLevel()) + tile.getColumn();
     }
 
     private long getTileNumber(TileKey tileKey)
     {
         return tileKey.getRow() < 0 ? -1 :
-            tileKey.getRow() * this.numColumnsInLevel(this.getLevel(tileKey.getLevelNumber())) + tileKey.getColumn();
+            (long) tileKey.getRow() * this.numColumnsInLevel(this.getLevel(tileKey.getLevelNumber())) + tileKey.getColumn();
     }
 
     /**
